agm: service: implement agm crash notificaiton to kernel
Add support in device object to write to an sysfs fd exposed
by the machine driver.Each time a device is opened, device object
writes into the sysfs node to notify the driver, similary when the
device is closed, device object will notify the same to the driver.
If agm service crashes before the device is closed, then the
driver would have an mismatch in the count suggesting an abrupt
crash in the agm userspace service.

Change-Id: I89c9979911a78894e3cb3c82046ca661720c92da