agm: client crash cleanup handling.
If two sessions are active and one of them gets closed,
then complete list of sessions is getting deleted,
resulting in other sessions not closing if client dies.

Now handling the case by adding a check which will
verify if session list is empty or not. If it is empty
then only the list is deleted.

Change-Id: Id20a2bb2398708a44cb50ec1ca04367de8e49020