agm: service: Set enable to sysfs node only on pcm_open success
If the sysfs node is updated before pcm_open and pcm_open fails, then
ASoC framework triggers dai_link shutdown that leads to a mismatch
in the sysfs node state and hence gets treated as an AGM crash, leading
to a CLEAN_ALL command to dsp.
Update Sysfs node with device enable only if pcm_open is successfull.

Change-Id: Idc0a9ce5f25be76d1dca38c97925f13b47cbbdc6