agm: Add aif state check while getting merged metadata
Since the aif object is not removed from the aif pool
of a session even after disconnecting aif and only
state gets changed to closed, stale aif
object metadata gets added in graph_add for next
session with the same handle. It leads to improper
GKV population. Check the aif obj state and ignore metadata
addition for closed objects to fix this.

Change-Id: I6c56ae4669a422ca7533e10a5c712d33ee3b6e0a