service: session: fix session state in session_prepare
If the session is already in started case and second backend
is connected,then session_prepare() incorrectly updates
the session state to prepared while it was in started state
previously. As a result next device switch fails as session 
remains in prepared state and graph_start is not called,
thus stream remains in muted state.

Change-Id: I48e67bf569ad38feab9adc51a80630a4a05eaf79